Move orbit correction inside the xtrack module
